+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added equivalent untis to `RexsStandardUnitIds` (`deg`, `degree`)
 
+### Fixed
+
+- RexsSchemaRegistry does now find REXS Versions also based on the additional model version strings
+
 
 ## [0.14.0] - 2025-07-18